% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLIMA.R
\name{Clima}
\alias{Clima}
\title{Automatizacion de la clasificacion climatica de Koppen modificada por Enriqueta Garcia}
\usage{
Clima(temperaturas, precipitaciones)
}
\arguments{
\item{temperaturas}{(vector) datos de la temperatura correspondiente a los doce meses y la temperatura anual}

\item{precipitaciones}{(vector) datos de la precipitacion correspondiente a los doce meses y la precipitacion anual}
}
\value{
una lista con el clima y cultivos
}
\description{
Regresa el, clima y cultivos que se pueden establecer en tal por su tipo de clima
}
